Move base64 and bytes tests with core
