+++ src/core/types/string/bytes.js
